backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.5.1
pydantic==2.7.4
pydantic-settings==2.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
